Update with 'move getConfig' changes
